razor pay and UI fixes
