+++ src/app/components/body/body.component.html
@@ -9,18 +9,13 @@
                 <span class="brand">Nora Tech</span>
                 <span class="example-spacer"></span>
 
-                <span class="menu">COMPANY</span>
-                <span class="menu">SERVICES</span>
-                <span class="menu">PRODUCTS</span>
-                <span class="menu" routerLink="/course">TRAINING</span>
-                <span class="menu" routerLink="/contact">CONTACT US</span>
-                <span class="menu">CAREER</span>
+                <a class="menu">COMPANY</a>
+                <a class="menu">SERVICES</a>
+                <a class="menu">PRODUCTS</a>
+                <a class="menu" routerLink="/course">TRAINING</a>
+                <a class="menu" routerLink="/contact">CONTACT US</a>
+                <a class="menu">CAREER</a>
 
-                <span class="menu">FEATURES</span>
-                <span class="menu">SOLUTIONS</span>
-                <a class="menu">SUCCESS STORIES</a>
-                  <a class="menu" routerLink="/contact">CONTACT US</a>
-                <a class="menu" routerLink="/course">COURSES</a>
 
                 <button mat-flat-button class="requestdemo">REQUEST DEMO</button>
                 <button mat-flat-button class="login" routerLink="/login">LOGIN</button>
@@ -33,15 +28,10 @@
               <a href="#">Company</a>
               <a href="#">Services</a>
               <a href="#">Products</a>
-              <a href="#">Trainings</a>
+              <a routerLink="/course">Trainings</a>
               <a href="#">Career</a>
               <a class="login" routerLink="/contact">Contact Us</a>
 
-              <span >FEATURES</span>
-              <span >SOLUTIONS</span>
-              <a >SUCCESS STORIES</a>
-              <a  routerLink="/contact">CONTACT US</a>
-              <a  routerLink="/course">COURSES</a>
 
             </div>
