+++ src/app/components/course-payment/enroll-student/enroll-student.component.html
@@ -41,7 +41,7 @@
         </mat-form-field>
         <br>
         <div>
-          <button mat-button matStepperNext>Next</button>
+          <button mat-button matStepperNext (click)="getOrderId()">Next</button>
         </div>
       </form>
     </mat-step>
